+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import FeedList from '@/app/ui/home/FeedList';
+import FeedList from '@/ui/home/FeedList';
 
 export default function Home() {
   return (
